Remove `AIQ` prefix from class and function names (#606)
## Description
* Strip the `AIQ` prefix from class names such that `AIQContext` =>
`Context`, the idea here is that we don't need the name of the library
in the class names as they are already in the package/namespace of the
library.
* Compatibility aliases were left in place to ease transition

## By Submitting this PR I confirm:
- I am familiar with the [Contributing
Guidelines](https://github.com/NVIDIA/NeMo-Agent-Toolkit/blob/develop/docs/source/resources/contributing.md).
- We require that all contributors "sign-off" on their commits. This
certifies that the contribution is your original work, or you have
rights to submit it under the same license, or a compatible license.
- Any contribution which contains commits that are not Signed-Off will
not be accepted.
- When the PR is ready for review, new or existing tests cover these
changes.
- When the PR is ready for review, the documentation is up to date with
these changes.

---------

Signed-off-by: David Gardner <dagardner@nvidia.com>
Signed-off-by: David Gardner <96306125+dagardner-nv@users.noreply.github.com>
Co-authored-by: Yuchen Zhang <134643420+yczhang-nv@users.noreply.github.com>